pkg/report: match FreeBSD KASAN reports
The format of the panic message means that we get lots of duplicate
reports.  Normalize KASAN report titles.  In particular, strip the code
name since it encodes the size of the redzone, which may vary depending
on the test case, especially if the report is triggered by a false
positive.